Separate UTF-codepage conversion into XUtil
